Added Travis CI config file

+++ Pod/Classes/BoredBoss.m
@@ -572,7 +572,7 @@ static BoredBoss *sharedInstance = nil;
         NSDictionary *json = [NSJSONSerialization JSONObjectWithData:responseData options:NSJSONReadingMutableContainers error:&error];
         
         if ([[json objectForKey:@"status"] isEqualToString:@"OK"]) {
-            BoredBossDebug(@"%@ %@ api saved %d items", self, endpoint, batchSize);
+            BoredBossDebug(@"%@ %@ api saved %ld items", self, endpoint, (unsigned long)batchSize);
         } else {
             BoredBossDebug(@"%@ %@ api rejected some items", self, endpoint);
         }
